ダイアログのstoreをuiから分離
